Create fiveg_n3 and fiveg_n4 libraries
Pin macaroonbakery to 1.3.2
The next protobuf version 4.21.0 has the breaking changes with the existing code
https://protobuf.dev/news/2022-05-06/#python-updates
