+++ test-requirements.txt
@@ -4,7 +4,7 @@ flake8-docstrings
 flake8-builtins
 isort
 juju
-macaroonbakery==1.3.2
+macaroonbakery==1.3.2  # https://protobuf.dev/news/2022-05-06/#python-updates
 mypy
 pep8-naming
 pyproject-flake8
