+++ tox.ini
@@ -10,7 +10,7 @@ envlist = lint, static, unit
 src_path = {toxinidir}/src/
 unit_test_path = {toxinidir}/tests/unit/
 integration_test_path = {toxinidir}/tests/integration/
-lib_path = {toxinidir}/lib/charms/sdcore_upf/v0/
+lib_path = {toxinidir}/lib/charms/sdcore_upf_k8s/v0/
 all_path = {[vars]src_path} {[vars]unit_test_path} {[vars]integration_test_path} {[vars]lib_path}
 
 [testenv]
